Add basic cli invocation for converter
So that we can run:

  gleam run -m html_lustre_converter/cli --target javascript --runtime deno -- file.html

And have the gleam code printed to standard out.

I am not familiar with the ins and outs of the best way to achieve
this. It might be valid to work with stdin (or at least support it
with a '-') rather than rely on a file. There might also be issues
with the new dependencies as the final plan to have a web interface
for the project.